Algoritam gotov
-problem konverzacije rijesen
-vizualno uljepsavanje metoda i klasa

+++ app/src/main/java/com/example/rene/navigation_on_a_ski_slope/ConvertingGpsCoordToXY.java
@@ -7,27 +7,27 @@ public class ConvertingGpsCoordToXY {
 
     public ConvertingGpsCoordToXY() {}
 
-    public float convertLon(float longitude) {  //pretvara long u nas sustav
+    public double convertLon(double longitude) {  //pretvara long u nas sustav
         lon = longitude;
         if(lon < 0){
             lon = lon * (-1);
             lon = lon/180;
-            return (float) lon;
+            return lon;
         }else {
             lon = lon/180;
-            return (float) lon;
+            return lon;
         }
     }
 
-    public float convertLat(float latitude) { //pretvara lat u nas sustav
+    public double convertLat(double latitude) { //pretvara lat u nas sustav
         lat = latitude;
         if(lat < 0){
             lat = lat * (-1);
             lat = lat/90;
-            return (float) lat;
+            return lat;
         }else {
             lat = lat/90;
-            return (float) lat;
+            return lat;
         }
     }
 }
